+++ RELEASE_NOTES.md
@@ -5,18 +5,19 @@ Please always read this file before installing the package
 
 Download software here: https://github.com/fritz-hh/scripts_NAS4Free/tags
 
-v2.4 (2023-10-07):
+vx.x (yyyy-mm-dd):
 ====
 
 New features
 ------------
 
-- None
+- manageAcpi.sh: Add an `-b` option to define that the NAS shall never go to sleep if an SMB connection to the NAS exists.
 
 Changes
 -------
 
 - Remove all trailing spaces in the scripts.
+- Use spaces for indentation instead of tabs (4 spaces per tab)
 
 Fixes
 -----